Get async home page test passing with update()
Our test of the successful login flow, which aims to test that when the
App renders, it checks for a logged-in user, then renders the Logged In
Home Page if one is present, the Public Home Page otherwise, was
stubbornly failing due to async issues.

This commit ensures that the render tree of the App is updated before we
expect() in the test.

madliberation-50