Pass props from App to Home Page
We want to pass the user from the App to the Logged In Home Page. We
may want to pass other props, like a function for changing state, from
the App to a Home Page.

The React Router docs
(https://reacttraining.com/react-router/web/guides/basic-components)
make it sound like you should use render= instead of component= when
passing props, or applying logic as we do when checking for a user, so
this switches us to using render=.

madliberation-50