Remove un-needed Lambda resource
It looked like the inclusion of Puppeteer was making the Lambda
resulting from the template too big. There was a failure event to the
effect of "zipped file too big" in the Cloud Formation stack details for
the madliberationjs-lambda stack.

We aren't using the Lambdas from this project. The api is in mljsapi.
I'm deleting the Get Hello World endpoint to see if that fixes it.

madliberation-90