Add a resource to the template
I am not using any resources from the Cloud Formation template in this
project. I just have it because it was part of the pipeline setup that
came with the CodeStar starter project. I removed the Lambda resource
from the starter template, because it was too large, probably due to the
inclusion of puppeteer.

This commit:
  1 - Adds a placeholder Resource to the template, because at least one
Resource is required in a Cloud Formation template.
  2 - Uses a troposphere script madliberationjs.py to generate the
template.

madliberation-90