Apply canary change to unit test
I'm making an intentionally breaking change to the home page. The
previous commit did that. The purpose is to test whether the build
properly deploys to staging, fails, and then doesn't deploy to prod. In
that previous commit, I forgot to allow for the breaking change in the
unit test.

madliberation-90