Display error, confirm join
From the Enter Room Code Page, you can now enter your room code and game
name, and on success it will take you to a page with a button and a
message to wait for the sedermaker, and on failure it displays a red
error message. On success it also tells you your Room Code and Game
Name.

madliberation-98