Replace nav with code, name, in app bar
This makes it so whenever a Room Code or Game Name is confirmed, they
are displayed in the App Bar. They wipe out the navigation menu,
because, once a Room Code or Game Name is confirmed, you should not be
navigating except under the control of the steps presented by the game.
For example, if you did Start Seder, and got a Room Code, then went to
Home and did Join a Seder, the seder you started could not be led by
anyone.

madliberation-98