Change API to api.passover.lol
A browser's "don't send cross-site cookies" setting makes the site not
work before this change. That's because we changed the front-end domain
to passover.lol, but left the API url as api.madliberationgame.com.

So the cookie from passover.lol (front end) to api.madliberationgame.com
(API) would be blocked.