Fix background image CSS
This looks good on all the devices available from Chrome dev tools, even
the one (iPad Pro) which previously showed a white stripe. It also seems
to fix the scrolling issue.

Also, it seems that if you zoom out you only shrink the buttons, which
is exactly what we want.

madliberation-155