Use visibilitychange to detect unload
When you minimize the browser or press the home button on mobile, many
browsers remove the page (ending my JavaScript) without firing
beforeunload or unload. Some recommendations say to use the Page
Visibility API, which, they say, is more likely to be available on
mobile as well as desktop. My site was demonstrably losing state without
intentional refresh on mobile Firefox.

This switches my concept of a page going away to follow the
visibilitychange event, to a state of document.hidden == true.

https://developers.google.com/web/updates/2018/07/page-lifecycle-api#legacy-lifecycle-apis-to-avoid
https://www.igvita.com/2015/11/20/dont-lose-user-and-app-state-use-page-visibility/

madliberation-168