Add eslint6, browserlist, try to fix formatting
https://github.com/douglasnaphas/madliberation/issues/191

Various formatting issues have come up. This fixes many of them. One
notable exception: the "submit all these" button on /play is always
visible, even when I can see it having the hidden property in the
Elements tab of dev tools. The next step to fix this would be to try
setting the visibility css property.