Remove unnecessary path props
YourRoomCodePage was not doing anything with chosenPath or
setChosenPath. The path (to the chosen script) gets recorded in the
database at the time the room code is generated, so there is no need for
the front-end to know about it after that.

There probably never will be a reason for the front-end to know about
the chosenPath after the room code is obtained. Even if I decide to
display the chosenPath later in the experience, once we have the concept
of users the state can be obtained from the server.