Do not reference before assignment
I was referencing a variable before assignment. This was reported as a
console warning. However, I cannot explain why it was not a compilation
error. I cannot explain the following series of outcomes in Node (in
particular, why the destructring assignment to x works but the one to m2
does not):

```
> const {m2 = { m2 }} = {x: 'y'}
Thrown:
ReferenceError: m2 is not defined
> x
Thrown:
ReferenceError: x is not defined
> const {x = { x }} = {x: 'y'}
undefined
> x
'y'
> const {x3 = { x2 }} = {x2: 'y'}
Thrown:
ReferenceError: x2 is not defined
```