Remove apparently unnecessary import of React
I thought this might cause problems, since in the past I've received
errors when not importing React. However, this works (the Generating
Room Code Page generates a spinner then generates a Room Code as
desired) in testing, so apparently the imports of React and Component
(which were reported in the console as warnings for this file but are
not for others) are unnecessary.